// types.ts
export interface GuideQuestion {
    id: string;
    text: string;
    type: 'text' | 'choice';
}

export interface GuideStep {
    title: string;
    description: string;
    questions: GuideQuestion[];
}